+++ include/reVec3d.h
@@ -41,6 +41,9 @@ namespace re
 		// Get vector magnitutde.
 		float length() const;
 
+		// Calculate absolute distance to another vector.
+		float distanceTo(const Vec3d& vector) const;
+
 		// Parallel vectors check.
 		bool isParallel(const Vec3d& vector) const;