+++ source/reMathUtil.cpp
@@ -10,7 +10,7 @@
 
 #include "reMathUtil.h"
 #include "reMath.h"
-#include <math.h>
+#include <cmath>
 
 float re::toRadians(float degrees)
 {
@@ -108,4 +108,13 @@ re::Matrix4 re::lookAt(const Vec3d& eye, const Vec3d& center, const Vec3d& up)
 	result.setTranslation(-eye);
 	
 	return result;
+}
+
+float re::triangleArea(const Vec3d & a, const Vec3d & b, const Vec3d & c)
+{
+	const float d1 = a.distanceTo(b);
+	const float d2 = a.distanceTo(c);
+	const float d3 = b.distanceTo(c);
+	const float halfPerimeter = (d1 + d2 + d3) / 2;
+	return sqrt(halfPerimeter * (halfPerimeter - d1) * (halfPerimeter - d2) * (halfPerimeter - d3));
 }